Fixed architecture.
