+++ src/com/acbelter/HashUtil.java
@@ -1,0 +1,31 @@
+package com.acbelter;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+public class HashUtil {
+    public static String generateHash(String data) {
+        if (data == null) {
+            return null;
+        }
+
+        MessageDigest md;
+        try {
+            md = MessageDigest.getInstance("SHA-256");
+        } catch (NoSuchAlgorithmException e) {
+            return null;
+        }
+        md.update(data.getBytes());
+
+        byte[] byteData = md.digest();
+        StringBuilder sb = new StringBuilder();
+        for (byte b : byteData) {
+            String hex = Integer.toHexString(0xff & b);
+            if (hex.length() == 1) {
+                sb.append('0');
+            }
+            sb.append(hex);
+        }
+        return sb.toString();
+    }
+}
