+++ src/com/acbelter/storage/HistoryFileStorage.java
@@ -1,0 +1,4 @@
+package com.acbelter.storage;
+
+public class HistoryFileStorage implements HistoryStorage {
+}
