+++ src/com/acbelter/storage/UserDataStorage.java
@@ -1,0 +1,13 @@
+package com.acbelter.storage;
+
+import com.acbelter.User;
+
+import java.io.IOException;
+
+public interface UserDataStorage {
+    void loadUserData() throws IOException;
+    boolean addUser(User user);
+    boolean updateUser(User user);
+    boolean isUserExists(String name);
+    User getUser(String name);
+}
